Financial_Forecasting
In this project we are going to perform a Finance Forecasting Tool which helps in predicting the future values of the previous data. This follows a recursive approach. Here we take the average of differences of all elements in the array and add that to the final element, thus finding the next probable element.